test(e2e): port hotfix bootstrap timing to the scenario registry
The merge from main brought in the e2e refactors that decoupled scenarios from
test control (70d6199c3e) and moved them to a standalone runner (a071b10dc5).
Scenario.T and RunScenario are both gone, so this scenario no longer compiled.

Register the two scenarios instead of declaring them as Test_ functions, and
move them out of the _test.go file, matching how every other scenario is now
declared. The parser unit test stays in _test.go.

Validators can no longer skip, so an absent hotfix now logs and passes rather
than skipping. That is the right outcome anyway: no hotfix configured means
there is nothing to time, which is not a failure of the code under test.
